Removes dependency on Medea package
Medea was primarily being used for it’s `AnyJSON` type’s ability to handle any JSON frament. But it’s not strictly legal (and has never been a good idea) to return anything but an object or array in response to a JSON request. And these types are easy to provide without Medea’s help via `JSONSerlialization` (and really, even this minimal amount of JSON parsing is less common now that Codable is a thing and decoding is happening at the callsite. Everyone just wants `Data`).

Of the rest of it’s features, the `ValidJSONObject` type is the only that was really being used, and that's trivial to implement internally to SessionArtist.

Breaking Changes:
* `Request.json(_:)` has been removed.
* `Medea` types like `JSONObject`, `JSONArray`, and `ValidJSONObject` are now directly under the `SessionArtist` namespace.
* The same events that threw errors that threw Medea-originated errors still throw errors. But now they're of the `SessionArtist.Error` type.
* Updated `Package` file to tools version 5.2. This should mean dependencies only used for testing (like `Perfidy`) will no longer be pulled by consumers.